feat: Implement Sidebar component
Created the Sidebar component responsible for rendering the sidebar of the application.
Added asset imports for various logos and images used in the sidebar.
The sidebar consists of a logo, a menu, and an account profile section.
Implemented the Menu component within the sidebar, which displays a list of menu items.
Used CurrentPageContext to track and update the current page displayed in the sidebar.
The MenuItem component represents each menu item with an image logo and text.